backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
openai==1.64.0
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyodbc==5.2.0
pytesseract==0.3.13
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
